requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
motor==3.3.1
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.3
numpy==1.26.2
python-magic==0.4.27
pefile==2023.2.7
yara-python==4.3.1
virustotal-api==1.1.11
celery==5.3.4
redis==5.0.1 